move assert slashedOnChain into `assert.dispute`
